demo alternating on reboots

+++ demo.sh
@@ -8,12 +8,23 @@ sudo killall -9 rpitx
 sudo killall -9 sendiq
 sudo fuser -k 8080/tcp
 
-while true; do
-      sleep 1;
-      echo -e "\n Changing mode ********************************************************************************\n\n"
+y=$(last reboot | grep ^reboot | wc -l)
+echo $y
+
+if (( $y % 2 == 0 )) 
+then
+	echo -e "\n Continuous FSK Mode\n\n"
+	/home/pi/CubeSatSim/radioafsk f 7
+else
+
+	echo -e "\n Alternating FSK, BPSK, and AFSK telemetry Mode \n\n"
+	while true; do
+      		sleep 1;
+      		echo -e "\n Changing mode ********************************************************************************\n\n"
       
-      timeout 25 /home/pi/CubeSatSim/radioafsk c 3
-      sudo killall -9 rpitx
-      sudo killall -9 sendiq
-      sudo fuser -k 8080/tcp
-done
+      		timeout 25 /home/pi/CubeSatSim/radioafsk c 3
+      		sudo killall -9 rpitx
+      		sudo killall -9 sendiq
+      		sudo fuser -k 8080/tcp
+	done
+fi
